[sandbox] Move sandbox info to podsandbox controller
